+++ src/Contracts/ApiGateway.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Ubient\PwnedPasswords\Api;
+namespace Ubient\PwnedPasswords\Contracts;
 
 interface ApiGateway
 {
@@ -8,7 +8,6 @@ interface ApiGateway
      * Indicates how frequently a password was found to be pwned.
      *
      * @param  string $password
-     * @throws \RuntimeException
      * @return int
      */
     public function search(string $password): int;