+++ src/PwnedPasswordsServiceProvider.php
@@ -2,11 +2,13 @@
 
 namespace Ubient\PwnedPasswords;
 
-use Ubient\PwnedPasswords\Rules\Pwned;
-use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Validator;
-use Ubient\PwnedPasswords\Api\ApiGateway;
+use Illuminate\Support\ServiceProvider;
 use Ubient\PwnedPasswords\Api\PwnedPasswordsGateway;
+use Ubient\PwnedPasswords\Contracts\ApiGateway;
+use Ubient\PwnedPasswords\Contracts\LookupErrorHandler;
+use Ubient\PwnedPasswords\Handler\LogHandler;
+use Ubient\PwnedPasswords\Rules\Pwned;
 
 class PwnedPasswordsServiceProvider extends ServiceProvider
 {
@@ -21,5 +23,6 @@ class PwnedPasswordsServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->bind(ApiGateway::class, PwnedPasswordsGateway::class);
+        $this->app->bind(LookupErrorHandler::class, LogHandler::class);
     }
 }